requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bytecode==0.17.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
ddtrace==4.1.0
envier==0.6.1
fastapi==0.125.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
httptools==0.7.1
idna==3.11
importlib_metadata==8.7.0
opentelemetry-api==1.39.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyparsing==3.2.5
python-dotenv==1.2.1
python-json-logger==4.0.0
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rsa==4.9.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.1
zipp==3.23.0
